requirements.txt
Flask
maxminddb
ip2location
werkzeug
gunicorn
git+https://github.com/lionsoul2014/ip2region.git#subdirectory=binding/python
ipip-ipdb
